refactor steps
