+++ composite-controller/generate-selector/deploy/case-2/controller.yaml
@@ -3,13 +3,9 @@ kind: CompositeController
 metadata:
   name: ping-pong-controller
 spec:
-  generateSelector: true
   parentResource:
     apiVersion: example.com/v1
     resource: pings
-    revisionHistory:
-      fieldPaths:
-        - spec.name
   childResources:
     - apiVersion: example.com/v1
       resource: pongs
