--- composite-controller/generate-selector/deploy/ping-match-expressions.yaml
@@ -1,14 +1,0 @@
-apiVersion: example.com/v1
-kind: Ping
-metadata:
-  name: shovan-maity
-  labels:
-    name: shovan-maity
-spec:
-  name: Shovan Maity
-  selector:
-    matchExpressions:
-    - key: name
-      operator: In
-      values:
-      - shovan-maity
