--- composite-controller/generate-selector/deploy/ping-match-labels.yaml
@@ -1,11 +1,0 @@
-apiVersion: example.com/v1
-kind: Ping
-metadata:
-  name: shovan-maity
-  labels:
-    name: shovan-maity
-spec:
-  name: Shovan Maity
-  selector:
-    matchLabels:
-      name: shovan-maity
