--- composite-controller/generate-selector/deploy/ping.yaml
@@ -1,8 +1,0 @@
-apiVersion: example.com/v1
-kind: Ping
-metadata:
-  name: shovan-maity
-  labels:
-    name: shovan-maity
-spec:
-  name: Shovan Maity
