+++ composite-controller/readme.md
@@ -1,7 +1,8 @@
   1. [Example of generate-selector.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/generate-selector)
-  2. [Example of revision-history.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/revision-history)
-  3. [Example of update-strategy-inplace.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-inplace)
-  4. [Example of update-strategy-ondelete.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-ondelete)
-  5. [Example of update-strategy-recreate.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-recreate)
-  6. [Example of update-strategy-rolling-recreate.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-rolling-recreate)
-  7. [Example of update-strategy-rolling-inplace.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-rolling-inplace)
+  2. [Example of spec-selector.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/spec-selector)
+  3. [Example of revision-history.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/revision-history)
+  4. [Example of update-strategy-inplace.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-inplace)
+  5. [Example of update-strategy-ondelete.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-ondelete)
+  6. [Example of update-strategy-recreate.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-recreate)
+  7. [Example of update-strategy-rolling-recreate.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-rolling-recreate)
+  8. [Example of update-strategy-rolling-inplace.](https://github.com/shovanmaity/metacontroller-by-example/tree/master/composite-controller/update-strategy-rolling-inplace)
