+++ composite-controller/revision-history/deploy/crd.yaml
@@ -10,6 +10,8 @@ spec:
     plural: pings
     singular: ping
   scope: Namespaced
+  subresources:
+    status: {}
 ---
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
@@ -23,3 +25,5 @@ spec:
     plural: pongs
     singular: pong
   scope: Namespaced
+  subresources:
+    status: {}
