+++ composite-controller/spec-selector/deploy/controller.yaml
@@ -7,10 +7,10 @@ spec:
     apiVersion: example.com/v1
     resource: pings
   childResources:
-  - apiVersion: example.com/v1
-    resource: pongs
-    updateStrategy:
-      method: InPlace
+    - apiVersion: example.com/v1
+      resource: pongs
+      updateStrategy:
+        method: InPlace
   hooks:
     sync:
       webhook:
